refactor: PR 리뷰 반영
